backend/requirements.txt
alembic
Authlib
fastapi
celery
redis
httpx
ipython
itsdangerous
Jinja2
psycopg2
pytest
requests
SQLAlchemy
uvicorn
passlib
bcrypt
sqlalchemy-utils
python-multipart
pyjwt
GeoAlchemy2
stripe
python-dateutil
pydantic>=1.2.0,<2.0.0
pytest-mock
requests-mock